# Optimize getDefinedMessage by deduping during initialization