@kares 2016-12-07
Excellent Ivo, this is certainly merge material.
Throwing in some discussion about revisiting `dedupMap = Collections.synchronizedMap(new WeakHashMap<RubyString, WeakReference<RubyString>>());` I believe this should be further improved by not using a locking map impl -> possibly replacing the cache with a concurrent map.

... would certainly not want to trade locking for less memory usage with frozen string literals. thoughts?

there seems to be a note on that already around `freezeAndDedupString` : 
```
     * Note that this cache does some sync against the Ruby instance. This
     * could cause contention under heavy concurrent load, so a reexamination
     * of this design might be warranted.
```