@ivoanjo 2016-12-07
Fully agreed @kares.

I had that jotted down on my notebook as something to be explored next, and it wouldn't even be the first time I'd write a concurrent hash map with weak keys (you need some extra mojo to remove the weak references after they get collected).

If you think that's the way to go, I'm willing to take a stab at it :)