@kares 2016-12-07
@ivoanjo ah right `WeakConcurrentMap` - probably the reason why its not there in the first place.
believe it makes sense (if its not too much work on top of the JDK concurrent map) ... at `org.jruby.util`

the sync was introduced in https://github.com/ivoanjo/jruby/commit/e88911e6813acc5f734e7d16447fe4d07a9e7b91 as a bug work-around, you should look at https://github.com/jruby/jruby/issues/3171 for details