@headius 2016-12-07
I pushed an additional improvement that caches the defined? messages inline. This avoids the round-trip through the runtime and the defined messages map entirely.

We still would like to de-bottleneck the dedup cache, of course, since it's used many other places.