@headius 2016-12-07
My numbers with the additional inline-caching tweak and invokedynamic enabled:

Before:
After:

We can further improve these once we improve defined? to cache its lookups (like method calls do).