@nirvdrum 2016-12-08
Wow. Apologies if that #3171 fix was too blunt. But it seems like an easy performance win to fix then.