@ivoanjo 2016-12-09
@headius if I may ask a probably-derp question about your approach :smile: 
It obviously is much better to just have a pointer for the ruby string containing the message, rather than going the roundabout way of getting it from the `EnumMap`.

When I was cooking the patch, I investigated a bit and the reason why there was a `freezeAndDedup` there in the first place, was, according to b3fb6e42f7c076377655fb9d7ef012a924ee59bb, so that "We need to dedup defined message strings so they all have same id", which makes sense to me.

But your approach always seems to create new `FrozenString` instances, so intuitively it would seem to me that this would not hold. But I tried experimenting with it, and it does hold, I still get the same `object_id` every time with your patch. So... why? I'm scratching my head a bit :)