@ivoanjo 2016-12-18
@kares I've been trying to see where `freezeAndDedupString` is still being used in order to devise a benchmark that can then be used to test an alternative approach to the current weakhashmap and it seems to me that after the change in this PR, there aren't any performance-critical uses of it anymore, so the added complexity may not be entirely worth it.

Am I missing some use case?