@kares 2016-12-18
@ivoanjo thought that the map is a cache for eventually all frozen strings (haven't looked into the details)