@ivoanjo 2016-12-19
@kares I don't think frozen strings are deduped by default, e.g.

```ruby
jruby-9.1.5.0 :001 > x = 97.chr.freeze
 => "a" 
jruby-9.1.5.0 :002 > y = 'a'.freeze
 => "a" 
jruby-9.1.5.0 :003 > x.object_id
 => 2002 
jruby-9.1.5.0 :004 > y.object_id
 => 2004 
```

I think they are only deduped when they are read in as literals:

```ruby
jruby-9.1.5.0 :001 > x = 'a'.freeze
 => "a" 
jruby-9.1.5.0 :002 > y = 'a'.freeze
 => "a" 
jruby-9.1.5.0 :003 > x.object_id
 => 2002 
jruby-9.1.5.0 :004 > y.object_id
 => 2002 
```

This is why I'm having a bit of a difficulty in coding a benchmark to test this feature.